--- .vscode/launch.json
@@ -1,14 +1,0 @@
-{
-  "configurations": [
-    {
-      "type": "java",
-      "name": "Spring Boot-FastfoodsimulatorApplication<fastfoodsimulator>",
-      "request": "launch",
-      "cwd": "${workspaceFolder}",
-      "console": "internalConsole",
-      "mainClass": "br.com.ufc.fastfoodsimulator.FastfoodsimulatorApplication",
-      "projectName": "fastfoodsimulator",
-      "args": ""
-    }
-  ]
-}
