--- fastfoodsimulator/src/main/java/br/com/ufc/fastfoodsimulator/FastfoodsimulatorApplication.java
@@ -1,13 +1,0 @@
-package br.com.ufc.fastfoodsimulator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FastfoodsimulatorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FastfoodsimulatorApplication.class, args);
-	}
-
-}
