--- fastfoodsimulator/src/main/java/br/com/ufc/fastfoodsimulator/controller/Teste.java
@@ -1,16 +1,0 @@
-package br.com.ufc.fastfoodsimulator.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("api/example")
-public class Teste {
-
-	@GetMapping("/hello-world")
-	public ResponseEntity<String> get() {
-		return ResponseEntity.ok("Hello World!");
-	}
-}
