--- fastfoodsimulator/src/main/java/br/com/ufc/fastfoodsimulator/repository/UserRepository.java
@@ -1,6 +1,0 @@
-package br.com.ufc.fastfoodsimulator.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface UserRepository {}
