--- fastfoodsimulator/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-## Database Properties
-spring.datasource.url = jdbc:mysql://localhost:3306/treinaweb?useSSL=false
-spring.datasource.username = root
-spring.datasource.password = root
-## Hibernate Properties
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
-spring.jpa.hibernate.ddl-auto = update
