--- fastfoodsimulator/src/test/java/br/com/ufc/fastfoodsimulator/FastfoodsimulatorApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.ufc.fastfoodsimulator;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FastfoodsimulatorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
