+++ src/main/java/com/ufc/fastfoodsimulator/FastFoodSimulatorApplication.java
@@ -1,0 +1,21 @@
+package com.ufc.fastfoodsimulator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@SpringBootApplication
+@RestController
+public class FastFoodSimulatorApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FastFoodSimulatorApplication.class, args);
+    }
+
+    @GetMapping("/hello")
+    public String hello(@RequestParam(value = "name", defaultValue = "mundo") String name){
+        return String.format("Olá, %s", name);
+    }
+}
