+++ src/test/java/com/ufc/fastfoodsimulator/FastFoodSimulatorApplicationTests.java
@@ -1,0 +1,13 @@
+package com.ufc.fastfoodsimulator;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FastFoodSimulatorApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
